Crate ferrous_chloride

Re-exports

pub use nom;

Modules

constants

iter
Iterator Types and implementations for data structures In general, you should not need to instantiate these types manually

parser

serde
Serializer and Deserializer for HCL

utils

value

Macros

inline_whitespace

map_err
map_err_str(IResult<I, O, u32>) -> IResult<I, O, Error>

map_err_str
map_err_str(IResult<I, O, u32>) -> IResult<I, O, Error>

take_till_match
Take bytes until the child parser succeeds

whitespace

Enums

Error
Error type for parsing

KeyValuePairs
A set of (Key, Value) pairs which can exist in a merged or unmerged variant

MergeBehaviour
Merge behaviour when parsing HCL Documents

OneOrMany
Either a single value, or many values

Value
Value in HCL

Traits

AsOwned
Type which has borrowed content which is able to be turned into an Owned version

Mergeable
Type is mergeable

ScalarLength
Has scalar length

Functions

from_str
Deserialize a type T from a provided HCL String

parse_reader
Parse a HCL string from a IO stream reader

parse_slice
Parse a HCL string from a slice of bytes

parse_str
Parse a HCL string into a [ConfigFile] which is close to an abstract syntax tree of the HCL string.