Enum ferrous_chloride::KeyValuePairs

pub enum KeyValuePairs<K, V> where
    K: Hash + Eq,  {
    Merged(StdHashMap<K, V, DefaultHashBuilder>),
    Unmerged(Vec<(K, V)>),
}

A set of (Key, Value) pairs which can exist in a merged or unmerged variant

A merged variant can only have unique keys, where the unmerged variant may have duplicate keys

Variants
Merged(StdHashMap<K, V, DefaultHashBuilder>)Unmerged(Vec<(K, V)>)
Methods

impl<K, V> KeyValuePairs<K, V> where
K: Hash + Eq,

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn unwrap_merged(self) -> StdHashMap<K, V, DefaultHashBuilder>

pub fn unwrap_unmerged(self) -> Vec<(K, V)>

Important traits for KeyValuePairsIterator<'a, K, V>
impl<'a, K: 'a, V: 'a> Iterator for KeyValuePairsIterator<'a, K, V> type Item = (&'a K, &'a V);

pub fn iter(&self) -> KeyValuePairsIterator<K, V>

Important traits for KeyIterator<'a, K, V>
impl<'a, K: 'a, V: 'a> Iterator for KeyIterator<'a, K, V> type Item = &'a K;

pub fn keys(&self) -> KeyIterator<K, V>

Important traits for ValueIterator<'a, K, V>
impl<'a, K: 'a, V: 'a> Iterator for ValueIterator<'a, K, V> type Item = &'a V;

pub fn values(&self) -> ValueIterator<K, V>

pub fn get_single<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Eq + Hash,

Get a single value with the specified key.

Warning

If the variant is unmerged, this operation will only return the first matching key it sees. A Vec's order might not be stable.

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<OneOrMany<&V>> where
K: Borrow<Q>,
Q: Eq + Hash,