Enum ferrous_chloride::parser::block::BlockBody

pub enum BlockBody<'a> {
    Body(Vec<Body<'a>>),
    Labels {
        empty: Vec<Body<'a>>,
        labels: HashMap<BlockLabel<'a>, BlockBody<'a>>,
    },
}

Variants
Body(Vec<Body<'a>>)Labels

Variant where there are block bodies with at least one label

You should not construct this enum by hand. It is a logic error for the labels HashMap to be empty and you can expect panics or unexpected behaviour from other functions in this library.

Fields of Labels

empty: Vec<Body<'a>>labels: HashMap<BlockLabel<'a>, BlockBody<'a>>

Methods

impl<'a> BlockBody<'a>

pub fn new() -> Self

pub fn append(&mut self, labels: Vec<BlockLabel<'a>>, body: Body<'a>)

pub fn get<S>(&self, labels: &[S]) -> Option<&Self> where
S: AsRef<str>,

pub fn get_mut<S>(&mut self, labels: &[S]) -> Option<&mut Self> where
S: AsRef<str>,

pub fn flat_iter<'b>(
&'b self
) -> Box<dyn Iterator<Item = (VecDeque<&'a str>, &'b Body<'a>)> + 'b> where
'b: 'a,

pub fn flat_iter_mut<'b>(
&'b mut self
) -> Box<dyn Iterator<Item = (VecDeque<&'a str>, &'b mut Body<'a>)> + 'b> where
'b: 'a,

pub fn flat_into_iter(
self
) -> Box<dyn Iterator<Item = (VecDeque<Cow<'a, str>>, Body<'a>)> + 'a>

pub fn len_blocks(&self) -> usize

pub fn has_further_labels(&self) -> bool

Indicates that there are further labels for the block

pub fn get_empty(&self) -> &[Body<'a>]

Borrow the bodies with no labels

pub fn get_labels(&self) -> Option<&HashMap<BlockLabel<'a>, BlockBody<'a>>>

Borrow the bodies with additional labels

pub fn label_lens(&self) -> BTreeSet<usize>

Get a set of label lengths for this body

pub fn label_lens_uniform(&self) -> Option<usize>

Return a uniform label lens, if any. This is cheaper to computer than label_lens

pub fn is_label_lens_uniform(&self) -> bool