Enum ferrous_chloride::parser::Expression

pub enum Expression<'a> {
    Null,
    Number(Number<'a>),
    Boolean(bool),
    String(Cow<'a, str>),
    Tuple(Tuple<'a>),
    Object(Object<'a>),
}

An Expression

Expression = (
    ExprTerm |
    Operation |  # Not supported
    Conditional # Not supported
);

ExprTerm = (
    LiteralValue |
    CollectionValue |
    TemplateExpr |
    VariableExpr |
    FunctionCall |
    ForExpr |
    ExprTerm Index |
    ExprTerm GetAttr |
    ExprTerm Splat |
    "(" Expression ")"
);

LiteralValue = (
  NumericLit |
  "true" |
  "false" |
  "null"
);

Numeric literals represent values of type number.

Variants
Null

A null HCL expression, expressed literally

Number(Number<'a>)

An arbitrary precision number

Boolean(bool)

A boolean value, expressed as the literals true or false

String(Cow<'a, str>)

A HCL string

Tuple(Tuple<'a>)

A HCL tuple (list)

Object(Object<'a>)

A HCL object (map)

Methods

impl<'a> Expression<'a>

pub fn parse(s: &'a str) -> Result<Self, Error>

Parse a string as a HCL expression

The string is expected to be fully consumed during parsing or an eror will be returned.

In general, this method should not be used. Prefer to use parse_str to parse a HCL configuration file instead.

pub fn new_tuple<T>(iterator: T) -> Self where
T: IntoIterator<Item = Expression<'a>>,

Convenience method to create a new Tuple Expression variant from an iterator of Expressions

pub fn new_object<T, I>(iterator: T) -> Self where
T: IntoIterator<Item = (I, Expression<'a>)>,
I: Into<ObjectElementIdentifier<'a>> + 'a,

Convenient method to create a new Object Expression variant from an iterator of Key-value pairs.

pub fn variant_name(&self) -> &'static str

Get the name of the Expression variant as a string.