Module ferrous_chloride::parser::identifier

Identifiers

Reference

Identifiers name entities such as blocks, attributes and expression variables. Identifiers are interpreted as per UAX #31 Section 2. Specifically, their syntax is defined in terms of the ID_Start and ID_Continue character properties as follows:

Identifier = ID_Start (ID_Continue | '-')*;

The Unicode specification provides the normative requirements for identifier parsing. Non-normatively, the spirit of this specification is that ID_Start consists of Unicode letter and certain unambiguous punctuation tokens, while ID_Continue augments that set with Unicode digits, combining marks, etc.

The dash character - is additionally allowed in identifiers, even though that is not part of the unicode ID_Continue definition. This is to allow attribute names and block type names to contain dashes, although underscores as word separators are considered the idiomatic usage.

Functions

identifier

Type Definitions

Identifier