Module ferrous_chloride::parser::whitespace

Whitespace and comments related

Reference

Comments and Whitespace are recognized as lexical elements but are ignored except as described below.

Whitespace is defined as a sequence of zero or more space characters (U+0020). Newline sequences (either U+000A or U+000D followed by U+000A) are not considered whitespace but are ignored as such in certain contexts.

Horizontal tab characters (U+0009) are not considered to be whitespace and are not valid within HCL native syntax.

Comments serve as program documentation and come in two forms:

Line comments start with either the // or # sequences and end with the next newline sequence. A line comments is considered equivalent to a newline sequence.

Inline comments start with the /* sequence and end with the */ sequence, and may have any characters within except the ending sequence. An inline comments is considered equivalent to a whitespace sequence.

Comments and whitespace cannot begin within within other comments, or within template literals except inside an interpolation sequence or template directive.

Functions

hash_comment

inline_comment

inline_whitespace

line_comment

newline

slash_comment

whitespace