Trait ferrous_chloride::Mergeable

pub trait Mergeable {
    fn is_merged(&self) -> bool;

    fn is_unmerged(&self) -> bool { ... }
}

Type is mergeable

Required methods

fn is_merged(&self) -> bool

Recursively checks that self is merged

This method should return true if all values recursively are merged.

Note that this method might not be the complement of is_unmerged.

Loading content...
Provided methods

fn is_unmerged(&self) -> bool

Recursively checks that self is unmerged

This method should return true if all values recursively are unmerged.

Note that this method might not be the complement of is_merged.

Loading content...