Trait ferrous_chloride::utils::SafeIndexing

pub trait SafeIndexing: Slice<RangeFull> + Sized {
    type Iter: Iterator<Item = usize>;
    fn safe_indices(&self) -> Self::Iter;
}

Associated Types

type Iter: Iterator<Item = usize>

Loading content...
Required methods

fn safe_indices(&self) -> Self::Iter

Returns an iterator that will only ever yield "safe" indices that do not cause panics when slicing

For example, in a str, this iterator should never yield an index in the middle of a single unicode codepoint

Loading content...