Trait ferrous_chloride::utils::SliceBoundary

pub trait SliceBoundary<R>: Slice<R> + Sized where
    R: RangeBounds<usize>, {
    fn is_slice_boundary(&self, index: usize) -> bool;

    fn safe_slice(&self, range: R) -> Option<Self> { ... }
}

Required methods

fn is_slice_boundary(&self, index: usize) -> bool

Indicate if an index is the start and/or end of some boundary

For example, in an implementation for a &str, this might be start and/or end of a UTF-8 code point sequence (see str::is_char_boundary).

Loading content...
Provided methods

fn safe_slice(&self, range: R) -> Option<Self>

Safe slicing. The start and the end of the range (if bounded) must be at a boundary

If they are unsafe, then the implementation will return None.

Loading content...