Type Definition ferrous_chloride::value::Block

type Block<'a> = KeyValuePairs<Vec<String>, MapValues<'a>>;

Contains a list of HCL Blocks sharing the same identifier with one or more labels differentiating each Block from each other.

Methods

impl<'a> Block<'a>

pub fn new_merged<T, K, S>(iter: T) -> Result<Self, Error> where
T: IntoIterator<Item = (K, MapValues<'a>)>,
K: IntoIterator<Item = S>,
S: ToString,

pub fn new_unmerged<T, K, S>(iter: T) -> Self where
T: IntoIterator<Item = (K, MapValues<'a>)>,
K: IntoIterator<Item = S>,
S: ToString,

pub fn merge(self) -> Result<Self, Error>

pub fn as_merged(&self) -> Result<Self, Error>

pub fn unmerge(self) -> Self

pub fn as_unmerged(&self) -> Self

pub fn borrow_keys(&self) -> KeyValuePairs<Vec<&str>, &MapValues<'a>>

Borrow the keys as Vec<&str> for more ergonomic indexing.

Usage

use ferrous_chloride::parser::literals::Key;
use ferrous_chloride::value::*;

let block = Block::new_unmerged(vec![(
    vec!["instance", "an_instance"],
    MapValues::new_unmerged(vec![
        (Key::new_identifier("name"), Value::from("an_instance")),
        (Key::new_identifier("image"), Value::from("ubuntu:18.04")),
        (
            Key::new_identifier("user"),
            Value::Block(Block::new_unmerged(vec![(
                vec!["test"],
                MapValues::new_unmerged(vec![(
                    Key::new_identifier("root"),
                    Value::from(true),
                )]),
            )])),
        ),
    ]),
)]);
let block = block.merge().unwrap();
let instance = block
    .borrow_keys()
    .get::<[&str]>(&["instance", "an_instance"])
    .unwrap()
    .unwrap_one();

Motivation

A Block is implemented as [KeyValuePairs] with Vec<String> as keys. Behind the scenes, a merged [KeyValuePairs] is backed by a [HashMap].

Retrieving a key from a [HashMap] involves using the [HashMap::get] method which specifies that to lookup a key of type K, you may use any type Q that implements [std::borrow::Borrow]<K>.

Since Vec<T> only implements Borrow<[T]>, a Vec<String> only implements Borrow<[String]>.

The implication is that we cannot lookup the HashMap<Vec<String>, Value> with a list of &str.

Consider the following:

use std::collections::HashMap;
use std::borrow::Borrow;

let mut hashmap: HashMap<Vec<String>, usize> = HashMap::default();
let _ = hashmap.insert(vec!["a".to_string(), "b".to_string()], 123);

// Let's try to retrieve the value
// The following won't compile
// let _ = hashmap.get(&["a", "b"]);

// We have to use this...
let _ = hashmap.get::<[String]>(&["a".to_string(), "b".to_string()]);

As you can see, this is not ergonomic at all.

Fundamentally, this is because it's not zero-cost to convert a Vec<Stirng> into a &[&str]. See this question on StackOverflow.

Alternatives

The unstable raw entry API might help with this in the future.