Trait ferrous_chloride::AsOwned

pub trait AsOwned {
    type Output: 'static;
    fn as_owned(&self) -> Self::Output;
}

Type which has borrowed content which is able to be turned into an Owned version

In other words, this type should have some lifetime 'a that can be turned into 'static.

Associated Types

type Output: 'static

Loading content...
Required methods

fn as_owned(&self) -> Self::Output

Returns a variant of Self where nothing is borrowed.

Loading content...